WIP renaming remaining helpers in core plugin
