+++ app/controllers/application_controller.rb
@@ -18,9 +18,9 @@ class ApplicationController < ActionController::Base
 
   helper ShinyCMS::AdminAreaHelper
 
-  helper ShinyCMS::PluginHelper
+  helper ShinyCMS::PluginsHelper
   helper ShinyCMS::SidekiqHelper
-  helper ShinyCMS::ShinyUserHelper
+  helper ShinyCMS::UsersHelper
 
   # Prevent Blazer from unloading all of the ShinyCMS helpers
   def self.clear_helpers; end
