+++ plugins/ShinyCMS/app/helpers/shinycms/users_helper.rb
@@ -8,7 +8,7 @@
 
 module ShinyCMS
   # Helper methods for dealing with users
-  module ShinyUserHelper
+  module UsersHelper
     def current_user_can?( capability, category = :general )
       current_user&.can? capability, category
     end
@@ -22,7 +22,7 @@ module ShinyCMS
     end
 
     def user_profile_link( user = current_user )
-      return user.name unless Plugins.loaded? :ShinyProfiles
+      return user.name unless ShinyCMS::Plugins.loaded? :ShinyProfiles
 
       link_to user.name, shiny_profiles.profile_path( user.username )
     end
