+++ plugins/ShinyCMS/app/helpers/shinycms/votes_helper.rb
@@ -7,11 +7,11 @@
 # ShinyCMS is free software; you can redistribute it and/or modify it under the terms of the GPL (version 2 or later)
 
 module ShinyCMS
-  # Helper method for turning a ShinyWithVotes URL param back into a class name
-  module ShinyVotesHelper
+  # Helper method for turning a WithVotes URL param back into a class name
+  module VotesHelper
     include ActsAsTaggableOn::TagsHelper
 
-    def class_from_votable_url( url_param )
+    def class_from_vote_url( url_param )
       url_param.tr( '-', '/' ).classify
     end
   end
